+++ api/internal/domain/entities/organisation_domains.go
@@ -5,7 +5,7 @@ import "time"
 // OrganisationDomainDBModel contains domains associated with a given organisation.
 // These can be both allow listed and deny listed domains.
 type OrganisationDomainDBModel struct {
-	ID             int        `db:"id"`
+	ID             int64      `db:"id"`
 	UUID           string     `db:"uuid"`
 	CreatedAt      time.Time  `db:"created_at"`
 	UpdatedAt      time.Time  `db:"updated_at"`
@@ -13,5 +13,5 @@ type OrganisationDomainDBModel struct {
 	Domain         string     `db:"domain" json:"domain" binding:"required" validate:"required"`
 	IsValidated    bool       `db:"is_validated"`
 	AllowList      bool       `db:"allow_list"`
-	OrganisationID *int       `db:"organisation_id"`
+	OrganisationID *int64     `db:"organisation_id"`
 }
